added user resource , basic auth and authorisation(potentially will omit cast and reviewers resources)

+++ routes/reviews.js
@@ -1,6 +1,8 @@
 const Router = require('koa-router');
 const bodyParser = require('koa-bodyparser');
 const model = require('../models/reviews');
+const can = require('../permissions/reviews');
+const auth = require('../controllers/auth');
 const etag = require('etag');
 const {validateReview} = require('../controllers/validation');
 const router = Router({
@@ -8,10 +10,10 @@ const router = Router({
 });
 
 router.get('/', getAll);
-router.post('/', bodyParser(),validateReview, createReview);
+router.post('/', bodyParser(),validateReview,auth, createReview);
 router.get('/:id([0-9]{1,})', getById);
-router.put('/:id([0-9]{1,})', bodyParser(),validateReview, updateArticle);
-router.del('/:id([0-9]{1,})', deleteReview);
+router.put('/:id([0-9]{1,})', bodyParser(),validateReview,auth, updateArticle);
+router.del('/:id([0-9]{1,})',auth, deleteReview);
 
 async function getAll(ctx) {
 	console.log("pass")
@@ -54,34 +56,75 @@ async function getById(ctx) {
 
 async function createReview(ctx) {
 
-	const body = ctx.request.body;
+	const permission = can.create(ctx.state.user);
+	console.log(permission)
+	if (!permission.granted) {
+		ctx.status = 403;
+	} else {
+		const body = ctx.request.body;
 
-	let result = await model.add(body);
+		let result = await model.add(body);
 
-	if (result) {
+		if (result) {
 
-		ctx.status = 201;
+			ctx.status = 201;
 
-		ctx.body = {
-			ID: result.insertId
+			ctx.body = {
+				ID: result.insertId
+			}
+
 		}
-
 	}
 
 }
 
 async function updateArticle(ctx) {
     let id =  ctx.params.id;
-	let update = await model.updateReview(
-		id,
-		ctx.request.body.reviewerid,
-		ctx.request.body.rating
-	)
 	
+	const permission = can.update(ctx.state.user,ctx.state.user);
+	console.log(permission)
+	if (!permission.granted) {
+		ctx.status = 403;
+	} else {
+		
+		let update = await model.updateReview(
+			id,
+			ctx.request.body.reviewerid,
+			ctx.request.body.rating
+		)
+
+		if (update) {
+
+			ctx.status = 201;
+
+			ctx.body = {
+				message:"Record Updated"
+			}
+
+		}
+	}
 }
 
 async function deleteReview(ctx) {
     let id = ctx.params.id;
-	let reviews = await model.deleteReview(id);
+	const permission = can.delete(ctx.state.user,ctx.state.user);
+	console.log(permission)
+	if (!permission.granted) {
+		ctx.status = 403;
+	} else {
+		
+		
+		let reviews = await model.deleteReview(id);
+
+		if (reviews) {
+
+			ctx.status = 201;
+
+			ctx.body = {
+				message:"Record Deleted"
+			}
+
+		}
+	}
 }
 module.exports = router
